trimpots problem

+++ Software/Peripheral/trimpot_driver.h
@@ -17,15 +17,16 @@ public:
 	Trimpot(ADC_HandleTypeDef *hadc,uint32_t ADC_CHANNEL);
 	void config();
 
-	uint16_t read();
-	uint16_t getData() const;
+	float read();
+	float getData() const;
 	float getData(float min, float max) const;
 private:
 	static const uint16_t MIN_DATA_VALUE = 40;
 	static const uint16_t MAX_DATA_VALUE = 4000;
 	ADC_HandleTypeDef *hadc;
 	uint32_t ADC_CHANNEL;
-	uint16_t data;
+	// data is always between 0 and 1
+	float data;
 
 };
